Added support for non-integer weights

+++ source/edge.cc
@@ -1,6 +1,6 @@
 #include "edge.h"
 
-Edge::Edge(Vertex v1, Vertex v2, int weight): end1{v1}, end2{v2}, weight{weight} {}
+Edge::Edge(Vertex v1, Vertex v2, double weight): end1{v1}, end2{v2}, weight{weight} {}
 
 bool Edge::operator==(const Edge& e) const{
         return ((end1 == e.end1 && end2 == e.end2) ||
